update version 1.0.1

+++ database/migrations/2023_06_23_115326_create_accounts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('accounts', function (Blueprint $table) {
             $table->id();
             $table->integer('user_id');
-            $table->integer('balance');
+            $table->float('balance');
             $table->timestamps();
         });
     }
